Add Networking
- Add Standalone Asio as Engine Dependency
- Remove Standalone freetype Engine Dependency
- Add numerous networking functions to project settings
- Initialize Audio and Scripting system in OnAttach() instead of Application constructor
- Initialize networking context for OnSceneStart/Stop() in editor and runtime
- Add .env file for projects to manage server settings
- Serialize/Deserialize .env file through project loading setup in AssetManager
- Refactor Application Run() to allow for constant delta time in physics simulation and include std::chrono library usage
- Add API for skipping and adding additional update cycles in main/update thread
- Add ProcessEventQueue() to Application and Networking contexts to handle async events
- Add AsyncBusyTimer in Engine Core to allow for expensive and precise timers that run on their own thread
- Added numerous event types to handle networking code and async timer code
- Added client/server networking code into Engine/Networking
- Add network component stubs for later development in Scene/Components.h
- Add functionality inside RuntimeUI to select widget, set widget text color, set widget background color, and make a widget selectable.
- Update Pong to include Online Gameplay mode
- Add Server App for hosting
- Add OnPhysicsCollisionEnd event/pipeline for scripting
- Update Logging to use trace as default debug logger
- Replace KG_CORE_*** with KG_*** when logging
- Add CRC32 hashing functionality in Core/Filesystem.h
- Update AppTick Generator API to dynamically add/remove generators when needed
- Add Recurring Timer to AsyncBusyTimer API

+++ Engine/dependencies/msdf_atlas_gen/msdf-atlas-gen/RectanglePacker.cpp
@@ -53,15 +53,15 @@ int RectanglePacker::pack(Rectangle *rectangles, int count) {
             for (size_t j = 0; j < remainingRects.size(); ++j) {
                 const Rectangle &rect = rectangles[remainingRects[j]];
                 if (rect.w == space.w && rect.h == space.h) {
-                    bestSpace = i;
-                    bestRect = j;
+                    bestSpace = (int)i;
+                    bestRect = (int)j;
                     goto BEST_FIT_FOUND;
                 }
                 if (rect.w <= space.w && rect.h <= space.h) {
                     int fit = rateFit(rect.w, rect.h, space.w, space.h);
                     if (fit < bestFit) {
-                        bestSpace = i;
-                        bestRect = j;
+                        bestSpace = (int)i;
+                        bestRect = (int)j;
                         bestFit = fit;
                     }
                 }
@@ -93,22 +93,22 @@ int RectanglePacker::pack(OrientedRectangle *rectangles, int count) {
             for (size_t j = 0; j < remainingRects.size(); ++j) {
                 const OrientedRectangle &rect = rectangles[remainingRects[j]];
                 if (rect.w == space.w && rect.h == space.h) {
-                    bestSpace = i;
-                    bestRect = j;
+                    bestSpace = (int)i;
+                    bestRect = (int)j;
                     bestRotated = false;
                     goto BEST_FIT_FOUND;
                 }
                 if (rect.h == space.w && rect.w == space.h) {
-                    bestSpace = i;
-                    bestRect = j;
+                    bestSpace = (int)i;
+                    bestRect = (int)j;
                     bestRotated = true;
                     goto BEST_FIT_FOUND;
                 }
                 if (rect.w <= space.w && rect.h <= space.h) {
                     int fit = rateFit(rect.w, rect.h, space.w, space.h);
                     if (fit < bestFit) {
-                        bestSpace = i;
-                        bestRect = j;
+                        bestSpace = (int)i;
+                        bestRect = (int)j;
                         bestRotated = false;
                         bestFit = fit;
                     }
@@ -116,8 +116,8 @@ int RectanglePacker::pack(OrientedRectangle *rectangles, int count) {
                 if (rect.h <= space.w && rect.w <= space.h) {
                     int fit = rateFit(rect.h, rect.w, space.w, space.h);
                     if (fit < bestFit) {
-                        bestSpace = i;
-                        bestRect = j;
+                        bestSpace = (int)i;
+                        bestRect = (int)j;
                         bestRotated = true;
                         bestFit = fit;
                     }
